auth/requirements.txt
flask==2.0.1
redis==3.5.3
hvac==0.11.2
pyjwt==2.1.0
bcrypt==3.2.0
requests==2.26.0